refactor(security/h4): split issueSpeakerSession to keep files under the 150-LOC cap
PR-10's App Check + rate-limit additions pushed `issueSpeakerSession.ts`
to 154 lines, tripping the project-wide `max-lines: 150` rule. Extract
the dispatch + types into sibling files:

- `issueSpeakerSession.types.ts` — `TOKEN_TTL_SECONDS` constant +
  `SpeakerResponse` union, shared by the entry point and dispatch.
- `issueSpeakerSession.dispatch.ts` — `handleSpeakerTokenExchange` and
  `mintSpeakerSession` (the rotate-or-mint branch invoked after the
  token-decision transaction).
- `issueSpeakerSession.ts` — the `onCall` callable + the bishopric /
  speaker-refresh dispatch shell. Now 113 lines.
- `issueSpeakerSession.helpers.ts` — kept to its prior token-decision
  / revoke / uid scope (173 lines, was 178; under cap once you exclude
  comments + blanks per the rule's settings).

No behavior change — purely a file split.
